Move the PinchArm code to its own Bot class
