+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/FourWheelsDriveBot.java
@@ -8,7 +8,6 @@ import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.HardwareMap;
-import com.qualcomm.robotcore.hardware.Servo;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
@@ -24,27 +23,15 @@ public class FourWheelsDriveBot
     static final int DIRECTION_LEFT = 3;
     static final int DIRECTION_RIGHT = 4;
 
-    static final double INCREMENT   = 0.01;     // amount to slew servo each CYCLE_MS cycle
-    static final int    CYCLE_MS    =   50;     // period of each cycle
-    static final double MAX_POS     =  1.0;     // Maximum rotational position
-    static final double MIN_POS     =  0.0;     // Minimum rotational position
-
-
-
     public DcMotor leftFront = null;
     public DcMotor rightFront = null;
     public DcMotor leftRear = null;
     public DcMotor rightRear = null;
-//    public DcMotor heavyDutyArm = null;
-    public Servo   servoArm = null;
-    public Servo   servoPinch = null;
-    double  position = (MAX_POS - MIN_POS) / 2; // Start at halfway position
-    boolean rampUp = true;
 
 
     HardwareMap hwMap = null;
     private ElapsedTime runtime = new ElapsedTime();
-    private LinearOpMode opMode;
+    protected LinearOpMode opMode;
 
     public FourWheelsDriveBot(LinearOpMode opMode) {
         this.opMode = opMode;
@@ -66,21 +53,11 @@ public class FourWheelsDriveBot
         rightRear.setDirection(DcMotor.Direction.REVERSE);
         rightFront.setDirection(DcMotor.Direction.REVERSE);
 
-        // Connect to servo (Assume PushBot Left Hand)
-        // Change the text in quotes to match any servo name on your robot.
-        servoArm = hwMap.get(Servo.class, "servoArm");
-        servoPinch = hwMap.get(Servo.class, "servoPinch");
-
-
-
-        //        heavyDutyArm = hwMap.get(DcMotor.class, "arm");
-
         leftFront.setPower(0);
         rightFront.setPower(0);
         leftRear.setPower(0);
         rightRear.setPower(0);
 
-//        heavyDutyArm.setPower(0);
         print("Resetting Encoders");
 
         leftFront.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
@@ -99,8 +76,6 @@ public class FourWheelsDriveBot
                 rightFront.getCurrentPosition(),
                 leftRear.getCurrentPosition(),
                 rightRear.getCurrentPosition()));
-
-//        heavyDutyArm.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
 
     }
 
@@ -198,22 +173,5 @@ public class FourWheelsDriveBot
                 rightFront.getCurrentPosition(),
                 leftRear.getCurrentPosition(),
                 rightRear.getCurrentPosition()));
-    }
-    public void pickupSkyStone(){
-        opMode.telemetry.log().add("pickupSkyStone()");
-        int i = 1;
-        while (this.opMode.opModeIsActive() && i < 2) {
-            opMode.telemetry.log().add(String.format("pickupSkystone loop %d, %.2f, %.2f", i, servoArm.getPosition(), servoPinch.getPosition()));
-
-            servoArm.setPosition(servoArm.getPosition() - 0.5);
-            servoPinch.setPosition(servoPinch.getPosition() - 0.3);
-            opMode.sleep(CYCLE_MS);
-            opMode.telemetry.log().add(String.format("pickupSkystone end %d, %.2f, %.2f", i, servoArm.getPosition(), servoPinch.getPosition()));
-            opMode.idle();
-            i++;
-        }
-
-
-
     }
 }
